implemented create a request from a simulation, page to consult the status of the request, visor and animations for analyze the pdf's (torture), it will be helpful when evaluating, pending fix front and controller for evaluation(almost ready), the other mentioned things are ready at front and back

+++ PrestaBancoTingesoEv1/BackendPrestaBanco/src/main/java/com/example/prestabanco/services/SecondHomeReqService.java
@@ -16,8 +16,13 @@ public class SecondHomeReqService {
     private SecondHomeReqRepository secondHomeReqRepository;
     @Autowired
     private ClientRepository clientRepository;
-    public SecondHomeReqEntity createSecondHomeRequest(SimulationEntity simulation, byte[] appraisalCertificate,
-        byte[] incomeProof, byte[] jobContract, byte[] creditHistory, byte[] firstHomeDeed, BigDecimal monthlyIncome) {
+    public SecondHomeReqEntity createSecondHomeRequest(SimulationEntity simulation,
+                                                       byte[] appraisalCertificate,
+                                                       byte[] incomeProof,
+                                                       byte[] jobContract,
+                                                       byte[] creditHistory,
+                                                       byte[] firstHomeDeed,
+                                                       BigDecimal monthlyIncome) {
         int loanType = 2;
         SecondHomeReqEntity secondHomeRequest = new SecondHomeReqEntity();
 
@@ -32,7 +37,7 @@ public class SecondHomeReqService {
         secondHomeRequest.setLoanAmount(simulation.getLoanAmount());
         secondHomeRequest.setYears(simulation.getYears());
         secondHomeRequest.setCurrentStatus("En revisión inicial");
-        secondHomeRequest.setAnnualInterestRate(simulation.getAnnualInterestRate());
+        secondHomeRequest.setAnnualInterestRate(BigDecimal.valueOf(simulation.getAnnualInterestRate()));
         secondHomeRequest.setMonthlyPayment(simulation.getMonthlyPayment());
         secondHomeRequest.setMonthlyIncome(monthlyIncome);
         secondHomeRequest.setFirstHomeDeed(firstHomeDeed);
